match: +15 strict via r5 near-miss harvest wave (playbook lanes A-D)
16 in-worktree wins, composed A/B x2 = +15 gained 0 regressed, icf HONEST.
PropSync x4 (unsigned bounds-check cast), SHA1::Update (syntax-form CSE
defeat), DecodeDxtColor, MoggClip::SynthPoll (qualified devirt), MicManagerXbox
(file statics), Synapse::ProcessInPlace, CharEyes::EyesOnTarget,
BandIKEffector::ApplyConstraints, TrackWatcherImpl x2 (Min/Max fsel),
Game::E3Cheat, StreakMeter::SetPartColor, SongDB+GemManager (normalized wins,
strict blocked on Game.h) + CharIKFingers/UsbMidiGuitar banked fuzzy.

+++ src/system/bandobj/StreakMeter.cpp
@@ -165,7 +165,7 @@ void StreakMeter::SetPartColor(int i, VocalHUDColor col) {
         SetPartActive(i, false);
     else if (i < mPartColorAnims.size()) {
         RndPropAnim *curanim = mPartColorAnims[i];
-        if (curanim)
+        if ((int)curanim)
             curanim->SetFrame(col, 1.0f);
     }
 }
